// frontend/src/app/core/api/model/auditLogDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Represents a single immutable entry in the audit trail
 */
export interface AuditLogDto { 
    id?: string;
    requestId?: number;
    userId?: string;
    timestamp?: string;
    action?: string;
    previousStatus?: string;
    newStatus?: string;
    currentHash?: string;
    previousHash?: string;
}